restructuring - moving logic out of tests
