+++ perceptron/networks.py
